[Keystone] Don't sync DB, remove dependancy on non-existent module
This enables build of a Keystone image without a MySQL instance being
available.  Note that for a fresh deployment of Keystone it'll be
necessary to bootstrap by running the `db_sync` stuff as well as
creating the necessary identity service endpoints yourself.

Also removes specification of a module that's no longer required.