[Workflow] Fix image labelling
This minor change ensures that the labels defined in the Rockerfile
actually make it into the final image artefact.  As a bonus, you don't
end up with that mysterious dangling image and the end of every build.

It also improves the cleanup to use the 'prune' option native to the
'docker image' command, and removes the slightly dangerous removal of
all exited containers.