[Nova] Fix keystone_authtoken options for Liberty
The Mitaka branch of the module we're using doesn't quite set the right
selection of options in order for authentication via Keystone to be
successful.

Fix this by using the `nova_config` type directly to set the relevant
missing options.

Can probably be removed just as soon as we're on at least Mitaka.