[Workflow] Tag images with domain, minor cleanup, remove 'release'
This commit tags images with the domain for which they're built, thus
allowing us to push non-Production images into our Production registry
without causing conflict / confusion.

It also removes the 'release' that's part of the image tag as it's
arbitrary and often confusing.  This can still be optionally added to
the image metadata at time of build and inspected with somethign like:

```
$ docker inspect 85eaf31aa48f | jq '.[0] | .Config.Labels'
{
  "org.label-schema.build-date": "1705",
  "org.label-schema.name": "keystone",
  "org.label-schema.release": "mitaka",
  "org.label-schema.vcs-ref": "801da91",
  "org.label-schema.vcs-url": "https://github.com/datacentred/docker"
}
```

♻ and 🔧 - part of  PD-3016.