Refactor based on Puppet's official support for building Docker images
There's now an official `image_build` module which abstracts away the
Rube Goldberg melange of tools such as Rocker in order to simplify the
build process.

This commit refactors our build process to make use of this module and
removes the unmaintained `dcdcli` configuration (arguably this should
live in a seperate repo anyway).