Tag Keystone with version supporting signing_dir
Since this change:
https://github.com/openstack/puppet-keystone/commit/7dfc3cd0021b79e2be81da2ee4a5fc9954abc021
the `signing_dir` parameter is no longer available for Aodh (and
also for other modules not pointing to their master branch).

This commit pins puppet-keystone to the latest tag before the change. It
also replaces branch pinning with a tag pining for puppet-aodh and
removes unused aodh::auth class.

A long-term solution would be to bring Aodh's Puppet repo up-to-date,
but this require more testing.

Addresses: PD-3166